Make call an interface, and introduce Call.Factory.
More documentation forthcoming.
